load from installed gem files
