+++ .gitignore
@@ -7,3 +7,9 @@
 /pkg/
 /spec/reports/
 /tmp/
+
+# rspec failure tracking
+.rspec_status
+
+# todo
+todo.md
